requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-extensions==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-nested-routers==0.92.5
Faker==5.8.0
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
signals==0.0.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
